requirements.txt
Django==5.1.2
djangorestframework==3.15.2
dj-database-url==2.2.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
GitPython==3.1.43
supervisor==4.2.5
requests==2.32.3
whitenoise==6.7.0
django-cors-headers==4.5.0
django-environ==0.11.2
cryptography==43.0.3
pycryptodome==3.21.0